+++ pkg/crd/crds/crds_test.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -32,6 +32,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
 	"github.com/arangodb/kube-arangodb/pkg/apis/ml"
 	"github.com/arangodb/kube-arangodb/pkg/apis/replication"
+	"github.com/arangodb/kube-arangodb/pkg/apis/scheduler"
 	"github.com/arangodb/kube-arangodb/pkg/apis/storage"
 )
 
@@ -73,6 +74,7 @@ func Test_CRD(t *testing.T) {
 		{ml.ArangoMLStorageCRDName, MLStorageDefinitionWithOptions},
 		{ml.ArangoMLCronJobCRDName, MLCronJobDefinitionWithOptions},
 		{ml.ArangoMLBatchJobCRDName, MLBatchJobDefinitionWithOptions},
+		{scheduler.ArangoProfileCRDName, SchedulerProfileDefinitionWithOptions},
 	}
 
 	for _, tc := range testCases {
@@ -108,6 +110,7 @@ func Test_CRDGetters(t *testing.T) {
 		MLStorageWithOptions,
 		ReplicationDeploymentReplicationWithOptions,
 		StorageLocalStorageWithOptions,
+		SchedulerProfileWithOptions,
 	}
 	require.Equal(t, len(AllDefinitions()), len(getters))